TIP-842: Move family requirement model to Akeneo\Pim\Structure
